Leagues injector
